micro commit #1

+++ business-layer/src/main/java/com/ovchingus/service/Service.java
@@ -72,7 +72,7 @@ public class Service implements ServiceMethods {
     }
 
     @Override
-    public Integer buyProductsInOneStore(String storeName, String productName, Integer qty) {
+    public Double buyProductsInOneStore(String storeName, String productName, Integer qty) {
         return getSource().buyProductsInOneStore(storeName, productName, qty);
     }
